# openapi/artfct.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "artfct API",
    "version": "1.0.0",
    "description": "The data-plane API for encrypted ephemeral artifacts and permanent artifact bundles.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://artfct.dev"
    }
  ],
...
  "paths": {
...
    "/v1/artifacts": {
      "get": {
        "tags": ["Administration"],
        "summary": "List artifacts",
        "description": "Reserved admin listing endpoint.",
        "operationId": "listArtifacts",
        "x-status": "unimplemented",
        "security": [
          { "orgToken": [] },
          { "sessionJwt": [] }
        ],
        "responses": {
          "501": { "$ref": "#/components/responses/NotImplemented" }
        }
      },
      "post": {
        "tags": ["Artifacts"],
        "summary": "Create an artifact",
        "description": "Creates an anonymous encrypted artifact or starts an authenticated permanent bundle upload. Anonymous access is allowed only when mode is ephemeral.",
        "operationId": "createArtifact",
        "security": [
          {},
          { "orgToken": [] },
          { "sessionJwt": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateArtifactRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artifact created or permanent upload initialized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateArtifactResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "options": {
        "tags": ["Artifacts"],
        "summary": "Preflight artifact collection requests",
        "operationId": "optionsArtifacts",
        "security": [],
        "responses": {
          "204": { "$ref": "#/components/responses/Preflight" }
        }
      }
    },
...
    "/p/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ArtifactId" }
      ],
      "get": {
        "tags": ["Previews"],
        "summary": "Render an artifact preview",
        "operationId": "renderArtifact",
        "security": [],
        "responses": {
          "200": {
            "description": "HTML preview shell (ephemeral) or permanent artifact entrypoint on its isolated origin.",
            "headers": {
              "Content-Security-Policy": {
                "schema": { "type": "string" }
              },
              "X-Content-Type-Options": {
                "schema": { "type": "string", "const": "nosniff" }
              }
            },
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          },
          "302": {
            "description": "On the shared origin, the entrypoint of a live public permanent artifact redirects to the app's viewer (`{app}/a/{id}`, or `{app}/a/{id}/v/{n}` for `/p/{id}/v:{n}/`). Bundle subresources (`/p/{id}/{path}`) and ephemeral artifacts are served as before; isolated origins never redirect.",
            "headers": {
              "Location": { "schema": { "type": "string", "format": "uri" } }
            }
          },
          "404": {
            "description": "Artifact is missing, expired, incomplete, or inaccessible.",
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "options": {
        "tags": ["Previews"],
        "summary": "Preflight artifact preview requests",
        "operationId": "optionsArtifactPreview",
        "security": [],
        "responses": {
          "204": { "$ref": "#/components/responses/Preflight" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "orgToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "organization API token",
        "description": "Organization-scoped API token."
      },
      "sessionJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-TTL session JWT minted by the Laravel control plane."
      }
    },
    "parameters": {
      "ArtifactId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Artifact identifier: a 10-character ephemeral id, a 13-character lowercase base36 stable id (new permanent artifacts), or a 32-character hex id (permanent artifacts published before versioning). The stable id survives new versions.",
        "schema": {
          "type": "string",
          "pattern": "^(?:[A-Za-z0-9]{10}|[a-z0-9]{13}|[a-f0-9]{32})$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request could not be parsed.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required or invalid.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "The request exceeds a configured size limit.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The request is syntactically valid but violates the artifact contract.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected data-plane error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "NotImplemented": {
        "description": "This reserved operation is not implemented.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "not_implemented",
                "message": "This operation is not implemented.",
                "details": {}
              }
            }
          }
        }
      },
      "Preflight": {
        "description": "CORS preflight accepted. The response has no body.",
        "headers": {
          "Access-Control-Allow-Origin": {
            "schema": { "type": "string" }
          },
          "Access-Control-Allow-Methods": {
            "schema": { "type": "string" }
          },
          "Access-Control-Allow-Headers": {
            "schema": { "type": "string" }
          }
        }
      }
    },
    "schemas": {
...
      "CreateArtifactRequest": {
        "oneOf": [
          { "$ref": "#/components/schemas/EphemeralArtifactRequest" },
          { "$ref": "#/components/schemas/PermanentArtifactRequest" }
        ],
        "discriminator": {
          "propertyName": "mode",
          "mapping": {
            "ephemeral": "#/components/schemas/EphemeralArtifactRequest",
            "permanent": "#/components/schemas/PermanentArtifactRequest"
          }
        }
      },
      "EphemeralArtifactRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode",
          "body_ciphertext_b64",
          "body_iv_b64",
          "tier",
          "title",
          "description",
          "thumbnail",
          "preview_blurred",
          "provenance"
        ],
        "properties": {
          "mode": { "type": "string", "const": "ephemeral" },
          "body_ciphertext_b64": {
            "type": "string",
            "minLength": 1,
            "description": "AES-GCM ciphertext encoded as unpadded base64url."
          },
          "body_iv_b64": {
            "type": "string",
            "minLength": 1,
            "description": "A 12-byte AES-GCM nonce encoded as unpadded base64url."
          },
          "tier": { "$ref": "#/components/schemas/ArtifactTier" },
          "ttl_minutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 525600
          },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "thumbnail": { "type": "string", "format": "uri" },
          "preview_blurred": { "type": "boolean" },
          "provenance": { "$ref": "#/components/schemas/Provenance" }
        }
      },
      "PermanentArtifactRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode",
          "title",
          "description",
          "thumbnail",
          "preview_blurred",
          "manifest",
          "provenance"
        ],
        "properties": {
          "mode": { "type": "string", "const": "permanent" },
          "tier": { "$ref": "#/components/schemas/PermanentArtifactTier", "description": "Deprecated: use sharing. Ignored when sharing is given." },
          "sharing": { "$ref": "#/components/schemas/Sharing", "description": "Defaults to team when neither sharing nor tier is given. public is refused (403 public_sharing_disabled) when the team turned public sharing off." },
          "edit_access": { "$ref": "#/components/schemas/EditAccess", "description": "Defaults to view." },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "thumbnail": { "type": "string", "format": "uri" },
          "preview_blurred": { "type": "boolean" },
          "manifest": { "$ref": "#/components/schemas/Manifest" },
          "provenance": { "$ref": "#/components/schemas/Provenance" }
        }
      },
      "ArtifactTier": {
        "type": "string",
        "enum": ["public", "secure", "ephemeral"]
      },
      "PermanentArtifactTier": {
        "type": "string",
        "enum": ["public", "secure", "private"],
        "description": "Storage name of the sharing level, kept for compatibility: secure = team. New clients use `sharing`."
      },
      "Sharing": {
        "type": "string",
        "enum": ["private", "team", "public"],
        "description": "Who can open the artifact. private: the owner and team admins. team: any member of the owning team. public: anyone with the link."
      },
      "EditAccess": {
        "type": "string",
        "enum": ["view", "edit"],
        "description": "Whether the people the artifact is shared with may publish new versions. The owner can always edit. Public + edit still requires a signed-in Artfct account to publish."
      },
...
      "Provenance": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agent",
          "agent_raw",
          "agent_version",
          "model",
          "session_id",
          "tool",
          "repo_url",
          "branch",
          "commit_sha",
          "dirty",
          "source_path",
          "client",
          "client_version",
          "sources"
        ],
        "properties": {
          "agent": { "type": ["string", "null"] },
          "agent_raw": { "type": ["string", "null"] },
          "agent_version": { "type": ["string", "null"] },
          "model": { "type": ["string", "null"] },
          "session_id": { "type": ["string", "null"] },
          "tool": { "type": ["string", "null"] },
          "repo_url": { "type": ["string", "null"] },
          "branch": { "type": ["string", "null"] },
          "commit_sha": { "type": ["string", "null"] },
          "dirty": { "type": ["boolean", "null"] },
          "source_path": { "type": ["string", "null"] },
          "client": { "type": "string", "minLength": 1 },
          "client_version": { "type": "string", "minLength": 1 },
          "sources": { "$ref": "#/components/schemas/ProvenanceSources" }
        }
      },
...
      "CreateArtifactResponse": {
        "oneOf": [
          { "$ref": "#/components/schemas/EphemeralArtifactResponse" },
          { "$ref": "#/components/schemas/PermanentArtifactResponse" }
        ]
      },
      "EphemeralArtifactResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "url",
          "tier",
          "expires_at",
          "title",
          "description",
          "thumbnail",
          "preview_blurred"
        ],
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "tier": { "$ref": "#/components/schemas/ArtifactTier" },
          "expires_at": { "type": "string", "format": "date-time" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "thumbnail": { "type": "string", "format": "uri" },
          "preview_blurred": { "type": "boolean" }
        }
      },
      "PermanentArtifactResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "url", "tier", "sharing", "edit_access", "missing_files", "version"],
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "tier": { "$ref": "#/components/schemas/PermanentArtifactTier" },
          "sharing": { "$ref": "#/components/schemas/Sharing" },
          "edit_access": { "$ref": "#/components/schemas/EditAccess" },
          "version": { "type": "integer", "minimum": 1, "description": "The version this upload creates, or the existing current version when identical bytes were re-posted." },
          "missing_files": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "pattern": "^[a-f0-9]{64}$"
            }
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "details"],
            "properties": {
              "code": { "$ref": "#/components/schemas/ErrorCode" },
              "message": { "type": "string", "minLength": 1 },
              "details": { "type": "object" }
            }
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "invalid_json",
          "validation_failed",
          "invalid_artifact_id",
          "artifact_not_found",
          "body_too_large",
          "bundle_too_large",
          "entrypoint_missing",
          "invalid_path",
          "duplicate_path",
          "hash_mismatch",
          "file_count_exceeded",
          "unauthorized",
          "forbidden",
          "not_implemented",
          "internal_error",
          "authentication_required",
          "rate_limited",
          "quota_exceeded",
          "version_conflict",
          "version_not_found",
          "public_sharing_disabled"
        ]
      }
    }
  }
}